agent/requirements.txt
# LiveKit Agents
livekit-agents>=0.12.0
livekit-plugins-deepgram>=0.6.0
livekit-plugins-silero>=0.6.0
livekit-plugins-openai>=0.10.0

# Supabase
supabase>=2.0.0

# Environment
python-dotenv>=1.0.0
